+++ src/AppserverIo/Appserver/Application/Application.php
@@ -657,7 +657,19 @@ class Application extends \Thread implements ApplicationInterface
     public function registerClassLoaders()
     {
         foreach ($this->getClassLoaders() as $classLoader) {
+
+            // log the class loader we want to initialize
+            $this->getInitialContext()->getSystemLogger()->debug(
+                sprintf('Now register classloader %s for application %s', get_class($classLoader), $this->getName())
+            );
+
+            // register the class loader instance
             $classLoader->register(true, true);
+
+            // log the class loader we've successfully registered
+            $this->getInitialContext()->getSystemLogger()->debug(
+                sprintf('Successfully registered classloader %s for application %s', get_class($classLoader), $this->getName())
+            );
         }
     }
 
@@ -674,8 +686,18 @@ class Application extends \Thread implements ApplicationInterface
             // and IO read/write operations can propably cause segfaults!
             \Mutex::lock($this->mutex);
 
+            // log the manager we want to initialize
+            $this->getInitialContext()->getSystemLogger()->debug(
+                sprintf('Now register manager %s for application %s', get_class($manager), $this->getName())
+            );
+
             // initialize the manager instance
             $manager->initialize($this);
+
+            // log the manager we've successfully registered
+            $this->getInitialContext()->getSystemLogger()->debug(
+                sprintf('Now registered manager %s for application %s', get_class($manager), $this->getName())
+            );
 
             // unlock the mutex
             \Mutex::unlock($this->mutex);