+++ src/AppserverIo/Appserver/Core/Api/Node/DirectoryNodeInterface.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * AppserverIo\Appserver\Core\Api\Node\DirectoryNodeInterface
+ *
+ * PHP version 5
+ *
+ * @category   Server
+ * @package    Appserver
+ * @subpackage Application
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+
+namespace AppserverIo\Appserver\Core\Api\Node;
+
+use AppserverIo\Configuration\Interfaces\NodeValueInterface;
+
+/**
+ * Interface for the directory node information.
+ *
+ * @category   Server
+ * @package    Appserver
+ * @subpackage Application
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+interface DirectoryNodeInterface extends NodeValueInterface
+{
+
+    /**
+     * Returns the enforcement flag.
+     *
+     * @return boolean The enforcement flag
+     */
+    public function isEnforced();
+}