+++ src/AppserverIo/Appserver/Core/Api/Node/ScannerNode.php
@@ -1,0 +1,114 @@
+<?php
+
+/**
+ * AppserverIo\Appserver\Core\Api\Node\ScannerNode
+ *
+ * PHP version 5
+ *
+ * @category   Server
+ * @package    Appserver
+ * @subpackage Application
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+
+namespace AppserverIo\Appserver\Core\Api\Node;
+
+/**
+ * DTO to transfer the provisioner information.
+ *
+ * @category   Server
+ * @package    Appserver
+ * @subpackage Application
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+class ScannerNode extends AbstractNode implements ScannerNodeInterface
+{
+
+    /**
+     * A params node trait.
+     *
+     * @var \TraitInterface
+     */
+    use ParamsNodeTrait;
+
+    /**
+     * A directories node trait.
+     *
+     * @var \TraitInterface
+     */
+    use DirectoriesNodeTrait;
+
+    /**
+     * The scanner name.
+     *
+     * @var string
+     * @AS\Mapping(nodeType="string")
+     */
+    protected $name;
+
+    /**
+     * The scanner type.
+     *
+     * @var string
+     * @AS\Mapping(nodeType="string")
+     */
+    protected $type;
+
+    /**
+     * Initializes the scanner node with the necessary data.
+     *
+     * @param string $name        The scanner name
+     * @param string $type        The scanner type
+     * @param array  $params      The scanner params
+     * @param array  $directories The scanner directories
+     */
+    public function __construct($name = '', $type = '', array $params = array(), array $directories = array())
+    {
+
+        // initialize the UUID
+        $this->setUuid($this->newUuid());
+
+        // set the data
+        $this->name = $name;
+        $this->type = $type;
+        $this->params = $params;
+        $this->directories = $directories;
+    }
+
+    /**
+     * Returns the nodes primary key, the name by default.
+     *
+     * @return string The nodes primary key
+     * @see \AppserverIo\Appserver\Core\Api\Node\AbstractNode::getPrimaryKey()
+     */
+    public function getPrimaryKey()
+    {
+        return $this->getName();
+    }
+
+    /**
+     * Returns the provisioner type.
+     *
+     * @return string The provisioner type
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * Returns the provisioner name.
+     *
+     * @return string The provisioner name
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+}